% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ptd_icon.R
\name{geom_ptd_icon}
\alias{geom_ptd_icon}
\title{PTD Icons}
\usage{
geom_ptd_icon(
  data = NULL,
  icons_size = 8L,
  icons_position = c("top right", "bottom right", "bottom left", "top left"),
  ...
)
}
\arguments{
\item{data}{the dataframe to plot. Must be of type ptd_spc_df}

\item{icons_size}{the size of the icons}

\item{icons_position}{the position of the icons in the plot}

\item{...}{currently unused}
}
\description{
Inserts the Making Data Count icons for variation/assurance to a plot
}
