% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ptd_icon.R
\docType{data}
\name{GeomPTDIcon}
\alias{GeomPTDIcon}
\title{GeomPTDIcon}
\format{
An object of class \code{GeomPTDIcon} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomPTDIcon
}
\description{
The Geom for the PTD icons. See \code{geom_ptd_icon()}.
}
\keyword{datasets}
