% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex-logo.R
\name{hex_logo}
\alias{hex_logo}
\title{Generate hexagonal logo for the package}
\usage{
hex_logo(
  seed = 123,
  n_points = 10,
  fill_palette = c("#ED6571", "#F69489", "#FFCF86", "#91CBD7", "#4E9BB9"),
  fill_palette_breaks = seq(0.2, 0.8, length.out = length(fill_palette) - 1) * n_points,
  line_size = 0.5,
  point_size = 2.9,
  point_shape = 21,
  point_stroke = line_size,
  package = "NHSRplotthedots",
  main_colour = "#000000",
  background_colour = "#FFFFFF",
  p_x = 1,
  p_y = 1.62,
  p_size = 16,
  s_x = 1,
  s_y = 0.98,
  s_height = 1.17,
  s_width = 1.5,
  asp = 0.9,
  url = "https://github.com/nhs-r-community/NHSRplotthedots",
  u_angle = 30,
  u_x = 1.02,
  u_y = 0.08,
  u_size = 4.85,
  dpi = 500,
  out_filename = "inst/images/logo.png"
)
}
\arguments{
\item{seed}{numeric value for reproducibility}

\item{n_points}{numeric value with the number of points to generate}

\item{fill_palette}{vector of colours to fill the points}

\item{fill_palette_breaks}{numeric vector with the breaking points (to change
the colours of the points)}

\item{line_size}{numeric value with the line size/thickness}

\item{point_size}{numeric value with the size of the points}

\item{point_shape}{numeric value with the shape of the points}

\item{point_stroke}{numeric value with the size of the stroke for the points}

\item{package}{string with the package name}

\item{main_colour}{string with the main colour (text, url and border)}

\item{background_colour}{string with the background colour}

\item{p_x}{x position for the package name}

\item{p_y}{y position for the package name}

\item{p_size}{font size for package name}

\item{s_x}{x position for subplot}

\item{s_y}{y position for subplot}

\item{s_height}{height for subplot}

\item{s_width}{width for subplot}

\item{asp}{aspect ratio, only works if subplot is an image file}

\item{url}{package repository URL (or other relevant URL)}

\item{u_angle}{angle for URL}

\item{u_x}{x position for URL}

\item{u_y}{y position for URL}

\item{u_size}{text size for URL}

\item{dpi}{plot resolution}

\item{out_filename}{filename to save sticker}
}
\value{
sticker object (\code{gg} class)
}
\description{
Generate hexagonal logo for the package
}
\examples{
\dontrun{
# Simple plot with 10 points and 5 colours
hex_logo()

# Multiple combination of parameters
plots <- tibble::tribble(
  ~n_points, ~point_stroke, ~fill_palette,
  10, 0, c("#ED6571", "#F69489", "#FFCF86", "#91CBD7", "#4E9BB9"),
  10, 0, c("#ED6571", rep("#FFCF86", 3), "#4E9BB9"),
  10, .5, c("#ED6571", "#F69489", "#FFCF86", "#91CBD7", "#4E9BB9"),
  10, .5, c("#ED6571", rep("#FFCF86", 3), "#4E9BB9"),
  15, 0, c("#ED6571", "#F69489", "#FFCF86", "#91CBD7", "#4E9BB9"),
  15, 0, c("#ED6571", rep("#FFCF86", 3), "#4E9BB9"),
  15, .5, c("#ED6571", "#F69489", "#FFCF86", "#91CBD7", "#4E9BB9"),
  15, .5, c("#ED6571", rep("#FFCF86", 3), "#4E9BB9"),
  20, 0, c("#ED6571", "#F69489", "#FFCF86", "#91CBD7", "#4E9BB9"),
  20, 0, c("#ED6571", rep("#FFCF86", 3), "#4E9BB9"),
  20, .5, c("#ED6571", "#F69489", "#FFCF86", "#91CBD7", "#4E9BB9"),
  20, .5, c("#ED6571", rep("#FFCF86", 3), "#4E9BB9")
) \%>\%
  dplyr::rowwise() \%>\%
  dplyr::mutate(
    out_filename = stringr::str_c(
      "inst/images/logo_w_",
      n_points,
      "pt_",
      length(unique(fill_palette)),
      "col_",
      point_stroke,
      "st.png"
    )
  ) \%>\%
  purrr::pmap(hex_logo)
}
}
\keyword{internal}
