% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_function.R
\name{my_function}
\alias{my_function}
\title{This is an example function}
\usage{
my_function(input_a, input_b, colour = "blue")
}
\arguments{
\item{input_a}{A vector of values}

\item{input_b}{Another vector of values}

\item{colour}{A colour for the dots in the plot.  Has a default value of 'blue'.}
}
\value{
a ggplot2 object
}
\description{
The is to show you a little of what you can put in various sections for the package.
The function contains nothing important, but takes two input vectors, makes a data.frame then plots
it with `ggplot2`.  I chose this to show you how the 'import' comment works.  Use `roxygen2` package
to build the documentation and it will add this to both the description and the namespace.

The tags below are transcribed, by `roxygen2` into the correct slots for the help metadata.
}
\details{
ggplot2 returns a list but also plots it by default.
}
\examples{
a <- c(10,15,17,18,22)
b <- c(150, 220, 160, 190, 210)
my_function(a, b)

my_function(a, b, "red")

}
