% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ.R
\name{null-replacement}
\alias{null-replacement}
\alias{\%||\%}
\title{null replacement}
\usage{
a \%||\% b
}
\arguments{
\item{a}{a possibly null value}

\item{b}{a value to replace \code{a} with if \code{a} is null}
}
\description{
if \code{a} is null, choose \code{b}.
}
\examples{
\dontrun{
1 \%||\% 2 == 1
NULL \%||\% 2 == 2
}

}
