% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGgplot.R
\name{plot.ptd_spc_df}
\alias{plot.ptd_spc_df}
\title{Plot ptd_spc_df object}
\usage{
\method{plot}{ptd_spc_df}(x, ...)
}
\arguments{
\item{x}{data passed to .data argument of \code{\link[=create_spc_plot]{create_spc_plot()}}}

\item{...}{other arguments passed to \code{\link[=create_spc_plot]{create_spc_plot()}}}
}
\description{
Plot function for a ptd_spc_df object. It calls \code{\link[=create_spc_plot]{create_spc_plot()}}.
}
\seealso{
create_spc_plot
}
