% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptd_spc_colours.R
\name{ptd_spcColours}
\alias{ptd_spcColours}
\title{SPC Colours}
\usage{
ptd_spcColours(
  common_cause = "#7B7D7D",
  special_cause_improvement = "#289de0",
  special_cause_concern = "#fab428",
  value_line = "#7B7D7D",
  mean_line = "#000000",
  lpl = "#7B7D7D",
  upl = "#7B7D7D",
  target = "#361475",
  trajectory = "#de1b1b"
)
}
\arguments{
\item{common_cause, special_cause_improvement, special_cause_concern}{the colour of the points}

\item{value_line}{the colour of the line joining the points}

\item{mean_line}{the colour of the "mean" (average) line}

\item{lpl, upl}{the colour the the lower and upper process limit lines}

\item{target}{the colour of the target line}

\item{trajectory}{the colour of the trajectory line}
}
\value{
a list of colours
}
\description{
Produces a list of colours that controls the geoms in the plot
}
