% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptd_target.R
\name{ptd_target}
\alias{ptd_target}
\title{Target}
\usage{
ptd_target(...)
}
\arguments{
\item{...}{either a single values, or, named values of targets. See examples.}
}
\value{
a list
}
\description{
Produces an object that can be used for adding Targets to an SPC chart. This method provides two different ways to
add a target:
\enumerate{
\item You can either provide a single value, which will apply the same target to every facet of an SPC
\item You can provide named values of targets, where the names correspond to the names of the facets, in order to
have different targets for each facet
}
}
