% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.r
\name{spc}
\alias{spc}
\title{SPC Plotting Function}
\usage{
spc(data.frame, valueField, dateField, facetField = NULL, options = NULL)
}
\arguments{
\item{data.frame}{A data frame containing a value field, a date field,
and a category field (if for faceting). There should be no gaps in the time series
for each category.}

\item{valueField}{Specify the field name which contains the value data, to be plotted on y axis.
Field name can be specified using non-standard evaluation (i.e. no quotation marks).}

\item{dateField}{Specify the field name which contains the date data, to be plotted on x axis.
Field name can be specified using non-standard evaluation (i.e. no quotation marks).}

\item{facetField}{Optional: Specify field name which contains a grouping/faceting variable. SPC logic will be applied to each group separately, with outputs combined. Currently accepts 1 variable only.
Field name can be specified using non-standard evaluation (i.e. no quotation marks).}

\item{options}{Optional: A list object containing additional control and formatting properties. Preferably created using the spcOptions function.}
}
\description{
`spc` returns a plot object or data table with SPC values using NHSI 'plot the dots' logic.
}
\details{
This function is designed to produce consistent SPC charts
across Information Department reporting, according to the 'plot the dots'
logic produced by NHSI. The function can return either a plot or data frame.
}
