% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGgplot.R
\name{createGgplot}
\alias{createGgplot}
\alias{plot.ptd_spc_df}
\title{Create ggplot2}
\usage{
createGgplot(
  x,
  pointSize = 2.5,
  percentageYAxis = FALSE,
  mainTitle = "SPC Chart",
  xAxisLabel = NULL,
  yAxisLabel = NULL,
  fixedXAxisMultiple = TRUE,
  fixedYAxisMultiple = TRUE,
  xAxisDateFormat = "\%d/\%m/\%Y",
  xAxisBreaks = NULL,
  yAxisBreaks = NULL,
  ...
)

\method{plot}{ptd_spc_df}(
  x,
  pointSize = 2.5,
  percentageYAxis = FALSE,
  mainTitle = "SPC Chart",
  xAxisLabel = NULL,
  yAxisLabel = NULL,
  fixedXAxisMultiple = TRUE,
  fixedYAxisMultiple = TRUE,
  xAxisDateFormat = "\%d/\%m/\%Y",
  xAxisBreaks = NULL,
  yAxisBreaks = NULL,
  ...
)
}
\arguments{
\item{x}{an object created by \code{\link[=spc]{spc()}}}

\item{pointSize}{Specify the plotting point size for the ggplot output. Default is 2.5.}

\item{percentageYAxis}{Specify whether the y axis values are percentages. Accepted values are TRUE for percentage y
axis, FALSE for integer y axis. Defaults to FALSE.}

\item{mainTitle}{Specify a character string value for the ggplot title.}

\item{xAxisLabel}{Specify a character string value for the x axis title.}

\item{yAxisLabel}{Specify a character string value for the y axis title.}

\item{fixedXAxisMultiple}{Specify whether, if producing a faceted spc, x axis should be fixed for all facet plots.
Accepted values are TRUE for fixed x axes or FALSE for individual x axes.}

\item{fixedYAxisMultiple}{Specify whether, if producing a faceted spc, y axis should be fixed for all facet plots.
Accepted values are TRUE for fixed y axes or FALSE for individual y axes.}

\item{xAxisDateFormat}{Specify how dates on the x axis should be displayed. Format should be provided
as a character string using 'd m Y' etc syntax.}

\item{xAxisBreaks}{Specify an interval value for breaks on the x axis. Value should be a character string expressing
interval length and type, e.g. "3 months", "7 days".}

\item{yAxisBreaks}{Specify an interval value for breaks on the y axis. Value should be a numeric vector of length 1,
either an integer for integer scales or a decimal value for percentage scales. This option is ignored if faceting
is in use.}

\item{...}{currently ignored}
}
\value{
The ggplot2 object
}
\description{
Creates a ggplot2 object using the parameters passed in.
}
