% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHSRplotthedots-package.R
\docType{package}
\name{NHSRplotthedots-package}
\alias{NHSRplotthedots}
\alias{NHSRplotthedots-package}
\title{NHSRplotthedots: Draw XmR Charts for NHS 'Making Data Count' Programme}
\description{
Provides tools for drawing Statistical Process Control (SPC) charts. This package supports the NHS 'Making Data Count' programme, and allows users to draw XmR charts, use change points and apply rules with summary indicators for when rules are breached.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nhsrplotthedots.nhsrcommunity.com}
  \item \url{https://github.com/nhs-r-community/NHSRplotthedots}
  \item Report bugs at \url{https://github.com/nhs-r-community/NHSRplotthedots/issues}
}

}
\author{
\strong{Maintainer}: Tom Smith \email{thomas.smith103@nhs.net}

Authors:
\itemize{
  \item Christopher Reading \email{christopher.reading1@nhs.net}
  \item Simon Wellesley-Miller \email{simon.wellesley-miller@nhs.net}
  \item Zoë Turner \email{Zoe.Turner2@nottshc.nhs.uk} (\href{https://orcid.org/0000-0003-1033-9158}{ORCID})
  \item Tom Jemmett \email{thomas.jemmett@nhs.net} (\href{https://orcid.org/0000-0002-6943-2990}{ORCID})
  \item Chris Mainey \email{c.mainey@nhs.net} (\href{https://orcid.org/0000-0002-3018-6171}{ORCID})
  \item John MacKintosh \email{john.mackintosh3@nhs.scot}
  \item Marcos Fabietti \email{marcos.fabietti@nhs.net}
  \item Fran Barton \email{francis.barton@nhs.net} (\href{https://orcid.org/0000-0002-5650-1176}{ORCID})
}

Other contributors:
\itemize{
  \item NHS-R community \email{nhs.rcommunity@nhs.net} [copyright holder]
}

}
\keyword{internal}
