% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHSRplotthedots-package.R
\docType{package}
\name{NHSRplotthedots-package}
\alias{NHSRplotthedots}
\alias{NHSRplotthedots-package}
\title{NHSRplotthedots: Draw XmR Charts for NHSE/I 'Making Data Count' Programme}
\description{
Provides tools for drawing Statistical Process Control (SPC) charts. This package supports the NHSE/I programme 'Making Data Count', and allows users to draw XmR charts, use change points and apply rules with summary indicators for when rules are breached.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nhs-r-community.github.io/NHSRplotthedots}
  \item \url{https://github.com/nhs-r-community/NHSRplotthedots}
  \item Report bugs at \url{https://nhs-r-community.github.io/NHSRplotthedots/issues}
}

}
\author{
\strong{Maintainer}: Christopher Reading \email{christopher.reading1@nhs.net}

Authors:
\itemize{
  \item Simon Wellesley-Miller \email{simon.wellesley-miller@nhs.net}
  \item Zoë Turner \email{Zoe.Turner2@nottshc.nhs.uk} (\href{https://orcid.org/0000-0003-1033-9158}{ORCID})
  \item Tom Jemmett \email{thomas.jemmett@nhs.net} (\href{https://orcid.org/0000-0002-6943-2990}{ORCID})
  \item Tom Smith \email{thomas.smith5@nuh.nhs.uk}
  \item Chris Mainey \email{c.mainey@nhs.net} (\href{https://orcid.org/0000-0002-3018-6171}{ORCID})
  \item John MacKintosh \email{john.mackintosh3@nhs.scot}
  \item Marcos Fabietti \email{marcos.fabietti@nuh.nhs.uk}
  \item Fran Barton \email{francis.barton@nhs.net} (\href{https://orcid.org/0000-0002-5650-1176}{ORCID})
}

Other contributors:
\itemize{
  \item NHS-R community \email{nhs.rcommunity@nhs.net} [copyright holder]
}

}
\keyword{internal}
