% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.r
\name{spc}
\alias{spc}
\title{SPC Plotting Function}
\usage{
spc(.data, valueField, dateField, facetField = NULL, options = NULL)
}
\arguments{
\item{.data}{A data frame containing a value field, a date field,
and a category field (if for faceting). There should be no gaps in the time series
for each category.}

\item{valueField}{Specify the field name which contains the value data, to be plotted on y axis.
Field name can be specified using non-standard evaluation (i.e. no quotation marks).}

\item{dateField}{Specify the field name which contains the date data, to be plotted on x axis.
Field name can be specified using non-standard evaluation (i.e. no quotation marks).}

\item{facetField}{Optional: Specify field name which contains a grouping/faceting variable. SPC logic will be applied
    to each group separately, with outputs combined. Currently accepts 1 variable only.
Field name can be specified using non-standard evaluation (i.e. no quotation marks).}

\item{options}{Optional: A list object containing additional control and formatting properties. Preferably created
using the spcOptions function.}
}
\value{
A ggplot2 object of the spc charts.  This will automatically print the plot, but can also be saved as an
    object if you want to manipulate it further.
}
\description{
`spc` returns a plot object or data table with SPC values using NHSI 'plot the dots' logic.
}
\details{
This function is designed to produce consistent SPC charts
across Information Department reporting, according to the 'plot the dots'
logic produced by NHSI. The function can return either a plot or data frame.
}
\examples{
library(NHSRdatasets)
data("ae_attendances")

# Pick a trust at random to look at their data for two years
trust1 <- subset(ae_attendances, org_code == "RJZ" & type == 1)

# Basic chart with improvement direction decreasing
spc(trust1,
  valueField = "breaches", dateField = "period",
  options = spcOptions(improvementDirection = "decrease")
)


# Pick a few trust, and plot individually using facet
# Also set the x-axis scale to vary for each and date groups to 3 months
orgs <- ae_attendances$org_code \%in\% c("RAS", "RJZ", "RR1", "RJC", "RQ1")
trusts4 <- subset(ae_attendances, orgs & type == 1)

spc(trusts4,
  valueField = "breaches", dateField = "period", facetField = "org_code",
  options = spcOptions(
    improvementDirection = "decrease",
    fixedYAxisMultiple = FALSE,
    xAxisBreaks = "3 months"
  )
)


# Save the first chart as an object this time then alter the ggplot theme
my_spc <- spc(trust1,
  valueField = "breaches", dateField = "period",
  options = spcOptions(improvementDirection = "decrease")
)
my_spc + ggplot2::theme_classic()
}
